feat(fe): split the code
