+++ apps/frontend/app/next-auth/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-import { authOptions } from '@/lib/auth'
+import { authOptions } from '@/lib/auth/authOptions'
 import NextAuth from 'next-auth'
 import type { NextRequest } from 'next/server'
